+++ tests/testthat/test-utils.R
@@ -9,6 +9,8 @@ toks <- tokens(corp, remove_punct = TRUE, remove_symbols = TRUE) %>%
     tokens_tolower()
 
 wov <- textmodel_word2vec(toks, dim = 50, iter = 10, min_count = 2, sample = 1)
+wov_nn <- textmodel_word2vec(toks, dim = 50, iter = 10, min_count = 2, sample = 1,
+                             normalize = FALSE)
 
 test_that("analogy works", {
     
@@ -108,6 +110,89 @@ test_that("similarity works", {
     expect_error(
         similarity(wov, c(1, -1), mode = "words"),
         "words must be named"
+    )
+})
+
+test_that("probability works", {
+    
+    prob1 <- probability(wov_nn, "us", mode = "values")
+    expect_true(is.matrix(prob1))
+    expect_identical(
+        dimnames(prob1),
+        list(names(wov_nn$frequency), "us")
+    )
+    
+    prob2 <- probability(wov_nn, c("us", "people"), mode = "values")
+    expect_true(is.matrix(prob2))
+    expect_identical(
+        dimnames(prob2),
+        list(names(wov_nn$frequency), c("us", "people"))
+    )
+    
+    prob3 <- probability(wov_nn, "us", mode = "words")
+    expect_true(is.matrix(prob3))
+    expect_identical(
+        prob3[1,],
+        c("us" = "let")
+    )
+    expect_identical(
+        dim(prob3),
+        c(length(wov_nn$frequency), 1L)
+    )
+    
+    prob4 <- probability(wov_nn, c("us", "people"), mode = "words")
+    expect_true(is.matrix(prob4))
+    expect_identical(
+        prob4[1,],
+        c("us" = "let", "people" = "american")
+    )
+    expect_identical(
+        dim(prob4),
+        c(length(wov_nn$frequency), 2L)
+    )
+    expect_warning(
+        probability(wov_nn, c("xx", "yyy", "us"), mode = "values"),
+        '"xx", "yyy" are not found'
+    )
+    expect_true(
+        suppressWarnings(
+            is.matrix(probability(wov_nn, c("xx", "yyy"), mode = "values"))
+        )
+    )
+    expect_warning(
+        probability(wov_nn, c("xx", "yyy", "us"), mode = "words"),
+        '"xx", "yyy" are not found'
+    )
+    expect_true(
+        suppressWarnings(
+            is.matrix(probability(wov_nn, c("xx", "yyy"), mode = "words"))
+        )
+    )
+    
+    prob5 <- probability(wov_nn, c("us" = 1, "people" = -1), mode = "values")
+    expect_equal(ncol(prob5), 1)
+    expect_true(is.matrix(prob5))
+    expect_identical(
+        dimnames(prob5),
+        list(names(wov_nn$frequency), NULL)
+    )
+    
+    prob6 <- probability(wov_nn, c("us" = 1, "people" = -1), mode = "words")
+    expect_equal(ncol(prob6), 1)
+    expect_true(is.matrix(prob6))
+    expect_identical(
+        dimnames(prob6),
+        NULL
+    )
+    
+    expect_error(
+        probability(wov_nn, c(1, -1), mode = "words"),
+        "words must be named"
+    )
+    
+    expect_error(
+        probability(wov, c(1, -1), mode = "words"),
+        "textmodel_wordvector must be trained with normalize = FALSE"
     )
 })